refactor: fix pr review
