+++ src/domain/types.ts
@@ -1,0 +1,2 @@
+/** siglens 앱에서 현재 활성화된 OAuth provider. */
+export type SupportedOAuthProvider = 'google' | 'kakao';
